Residual changes commit

+++ VSCADA/typedefs.h
@@ -31,9 +31,9 @@ typedef struct{
     int gpioPin;
     int val;
     double calConst;
-    int calVal;
+    double calVal;
     void calData(){
-        calVal = val*calConst;
+        calVal = (double)val*calConst;
     }
     void updateVal(int newVal)
       {
